GEOMESA-1648 Using shared index api code for z3 index
* Z3 now uses table-sharing along with other indices
