+++ geomesa-accumulo/geomesa-accumulo-datastore/src/main/scala/org/locationtech/geomesa/accumulo/iterators/Z2DensityIterator.scala
@@ -16,7 +16,7 @@ import org.apache.accumulo.core.client.IteratorSetting
 import org.apache.accumulo.core.data.{Key, Value}
 import org.apache.accumulo.core.iterators.{IteratorEnvironment, SortedKeyValueIterator}
 import org.geotools.factory.Hints
-import org.locationtech.geomesa.accumulo.index.Z2Index
+import org.locationtech.geomesa.accumulo.index.AccumuloFeatureIndex
 import org.locationtech.geomesa.accumulo.index.legacy.z2.Z2IndexV1
 import org.locationtech.geomesa.accumulo.iterators.KryoLazyDensityIterator.DensityResult
 import org.locationtech.geomesa.curve.Z2SFC
@@ -89,10 +89,11 @@ object Z2DensityIterator {
    * Creates an iterator config for the z2 density iterator
    */
   def configure(sft: SimpleFeatureType,
+                index: AccumuloFeatureIndex,
                 filter: Option[Filter],
                 hints: Hints,
                 priority: Int = KryoLazyDensityIterator.DEFAULT_PRIORITY): IteratorSetting = {
-    val is = KryoLazyDensityIterator.configure(sft, Z2Index, filter, hints, priority)
+    val is = KryoLazyDensityIterator.configure(sft, index, filter, hints, priority)
     is.setIteratorClass(classOf[Z2DensityIterator].getName)
     is.addOption(TableSharingKey, sft.isTableSharing.toString)
     is
